# internal/gateway/openapi.embed.yaml
openapi: 3.1.0
info:
  title: Telemed API Gateway
  description: |
    The single public edge for the Sri Lankan telemedicine platform. Every
    frontend (patient Flutter app, doctor Flutter app, admin Next.js web)
    talks only to this surface; it never calls a backend service directly.

    Authentication is a Keycloak-issued JWT (RS256/RS512/ES256), sent as
    `Authorization: Bearer <token>`. The gateway verifies the signature and
    claims once at the edge and forwards the verified identity to backend
    services as `X-Telemed-User-ID` / `X-Telemed-Roles`; those headers are
    gateway-internal and are stripped from any inbound client request before
    they can influence anything.

    All monetary amounts are integer cents (LKR unless noted). All
    timestamps are RFC 3339 UTC. Every response is wrapped in the platform
    envelope: `{"data": ...}` on success, `{"code","message","fields",
    "request_id"}` on error.
  version: "0.1.0"
  contact:
    name: telemed platform team
servers:
  - url: https://api.yourapp.lk
    description: production
  - url: http://localhost:8080
    description: local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued access token, 15 minute lifetime.
...
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        per_page: { type: integer, example: 20 }
        total: { type: integer, example: 134 }
        total_pages: { type: integer, example: 7 }
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code. See httpx.ErrorCode in the platform package.
          example: SLOT_UNAVAILABLE
        message:
          type: string
          example: this field is required
        fields:
          type: object
          additionalProperties: { type: string }
        request_id:
          type: string
          example: 3f9c2b7a-9b3e-4b8a-8f2a-9d6a1c7e4a10
...
    PrescriptionCreateRequest:
      type: object
      required: [appointment_id, drugs]
      properties:
        appointment_id: { type: string, format: uuid }
        drugs:
          type: array
          items:
            type: object
            required: [name, dosage, frequency, duration]
            properties:
              name: { type: string }
              dosage: { type: string }
              frequency: { type: string }
              duration: { type: string }
              notes: { type: string }
        patient_sex: { type: string, enum: [female, male, other], description: Printed on the PDF only; not stored }
        patient_weight_kg: { type: number, minimum: 0.5, maximum: 400, description: Printed on the PDF only; not stored }
        patient_allergies: { type: string, maxLength: 1000, description: Printed on the PDF only; not stored }
    Prescription:
      type: object
      properties:
        id: { type: string, format: uuid }
        appointment_id: { type: string, format: uuid }
        pdf_url: { type: string, format: uri }
        qr_verification_url: { type: string, format: uri }
    PrescriptionVerification:
      type: object
      properties:
        valid: { type: boolean }
        doctor_slmc: { type: string }
        issued_at: { type: string, format: date-time }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIError' }
    Forbidden:
      description: Authenticated but not permitted (wrong role, wrong IP, wrong origin).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIError' }
    NotFound:
      description: No route or resource matched.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIError' }
...
paths:
...
  /api/v1/prescriptions:
    get:
      tags: [records]
      summary: Look up a prescription by appointment
      description: >
        One prescription per appointment. Query `appointment_id` is required.
        The post-call patient summary polls this until the doctor issues the prescription.
      parameters:
        - name: appointment_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Prescription' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [records]
      summary: Issue a prescription
      description: >
        Requires role `doctor`. Allowed once the consultation has started,
        including while it is in progress. Generates a PDF + QR; slow path
        (upload timeout class, 60s).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrescriptionCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Prescription' } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/verify/prescriptions/{prescriptionID}:
    get:
      tags: [records]
      summary: Public prescription verification
      security: []
      description: >
        Public. Verifies a prescription by id + HMAC, matching the QR code
        printed on the PDF. The source docs describe this as living on a
        separate `verify.yourapp.lk` domain; this route is the gateway's
        equivalent surface, see the report.
      parameters:
        - name: prescriptionID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: h
          in: query
          required: true
          schema: { type: string }
          description: HMAC-SHA256(prescription_id + doctor_id + secret)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/PrescriptionVerification' } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/prescriptions/{prescriptionID}:
    get:
      tags: [records]
      summary: One prescription
      parameters:
        - name: prescriptionID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Prescription' } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/prescriptions/{prescriptionID}/pdf:
    get:
      tags: [records]
      summary: Rendered prescription PDF
      description: Rendered on demand, so this route carries the upload timeout class.
      parameters:
        - name: prescriptionID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Forbidden' }
...